backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
SQLAlchemy==2.0.36
psycopg[binary]==3.2.6
python-dotenv==1.0.1
